requirements.txt
asgiref==3.8.1
coverage==7.8.0
Django==5.2.1
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
sqlparse==0.5.3
